wip setting up importing LT inside another repo
