+++ bin/main.d.ts
@@ -1,0 +1,4 @@
+/**
+ *  Run end to end Tableland
+ **/
+export declare const main: () => Promise<void>;
