+++ bin/up.d.ts
@@ -1,5 +1,2 @@
 #!/usr/bin/env node
-/**
- *  Run end to end Tableland
- **/
 export {};
