+++ tableland.config.example.ts
@@ -1,0 +1,7 @@
+const config = {
+    validatorDir: "../go-tableland",
+    hardhatDir: "../evm-tableland",
+    verbose: true
+};
+
+export default config;
